refactoring: add index.ts files for every component and update importds accordingly.
