+++ frontend/src/components/AudioPanel/index.ts
@@ -1,0 +1,2 @@
+export { default as AudioPanel } from "./AudioPanel.vue";
+export * from "./AudioPanel";
