+++ frontend/src/components/CodePanel/index.ts
@@ -1,0 +1,2 @@
+export { default as CodePanel } from "./CodePanel.vue";
+export * from "./CodePanel";
