+++ frontend/src/components/ImagePanel/index.ts
@@ -1,0 +1,2 @@
+export { default as ImagePanel } from "./ImagePanel.vue";
+export * from "./ImagePanel";
