+++ frontend/src/components/SettingsPanel/index.ts
@@ -1,0 +1,2 @@
+export { default as SettingsPanel } from "./SettingsPanel.vue";
+export * from "./SettingsPanel";
