+++ frontend/src/components/SpritesPanel/index.ts
@@ -1,0 +1,2 @@
+export { default as SpritesPanel } from "./SpritesPanel.vue";
+export * from "./SpritesPanel";
