+++ frontend/src/components/StatusBanner/index.ts
@@ -1,0 +1,1 @@
+export { default as StatusBanner } from "./StatusBanner.vue";
