+++ frontend/src/components/TextPanel/TextPanel.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it, vi, beforeEach } from "vitest";
 import { mount, flushPromises } from "@vue/test-utils";
-import TextPanel from "./TextPanel.vue";
-import * as client from "../../api/client";
+import { TextPanel } from "@/components/TextPanel";
+import * as client from "@/api/client";
 
 vi.mock("../../api/client");
