+++ frontend/src/components/TextPanel/index.ts
@@ -1,0 +1,2 @@
+export { default as TextPanel } from "./TextPanel.vue";
+export * from "./TextPanel";
