+++ frontend/src/components/UnityProjectPanel/index.ts
@@ -1,0 +1,2 @@
+export { default as UnityProjectPanel } from "./UnityProjectPanel.vue";
+export * from "./UnityProjectPanel";
