+++ frontend/src/components/generic/BaseField/index.ts
@@ -1,0 +1,2 @@
+export { default as BaseField } from "./BaseField.vue";
+export * from "./BaseField";
