+++ frontend/src/components/generic/PromptInputSection/index.ts
@@ -1,0 +1,2 @@
+export { default as PromptInputSection } from "./PromptInputSection.vue";
+export * from "./PromptInputSection";
