+++ frontend/src/components/generic/SmartField/index.ts
@@ -1,0 +1,2 @@
+export { default as SmartField } from "./SmartField.vue";
+export * from "./SmartField";
